+++ zfit/models/convolution.py
@@ -8,10 +8,7 @@ import tensorflow as tf
 import tensorflow_addons as tfa
 import tensorflow_probability as tfp
 
-try:
-    from typing import Literal
-except ImportError:  # TODO(3.8): remove
-    from typing_extensions import Literal
+from typing import Literal
 
 import zfit.z.numpy as znp
 from .functor import BaseFunctor