+++ zfit/util/improved_sheather_jones.py
@@ -2,11 +2,11 @@
 
 import numpy as np
 import tensorflow as tf
-from tf_quant_finance.math import root_search
 
 from . import binning as binning_util
 from .. import z
 from ..settings import ztypes
+from . import root_search
 
 
 @z.function(wraps="tensor")
@@ -122,6 +122,7 @@ def _find_root(function, N, squared_integers, grid_data_dct2):
     >>> np.allclose(ans, 9.237610787616029e-05)
     True
     """
+
     # From the implementation by Botev, the original paper author
     # Rule of thumb of obtaining a feasible solution
     N2 = tf.math.maximum(